refactor: :recycle: Refactored controller and service to have better architecture
